+++ tests/Composer/Test/Autoload/AutoloadGeneratorTest.php
@@ -60,14 +60,14 @@ class AutoloadGeneratorTest extends TestCase
     protected function tearDown()
     {
         if ($this->vendorDir === $this->workingDir) {
-            if (is_dir($this->workingDir.'/.composer')) {
-                $this->fs->removeDirectory($this->workingDir.'/.composer');
+            if (is_dir($this->workingDir.'/composer')) {
+                $this->fs->removeDirectory($this->workingDir.'/composer');
             }
         } elseif (is_dir($this->vendorDir)) {
             $this->fs->removeDirectory($this->vendorDir);
         }
-        if (is_dir($this->workingDir.'/.composersrc')) {
-            $this->fs->removeDirectory($this->workingDir.'/.composersrc');
+        if (is_dir($this->workingDir.'/composersrc')) {
+            $this->fs->removeDirectory($this->workingDir.'/composersrc');
         }
 
         chdir($this->dir);
@@ -78,22 +78,22 @@ class AutoloadGeneratorTest extends TestCase
         $package = new MemoryPackage('a', '1.0', '1.0');
         $package->setAutoload(array(
             'psr-0' => array('Main' => 'src/', 'Lala' => array('src/', 'lib/')),
-            'classmap' => array('.composersrc/'),
+            'classmap' => array('composersrc/'),
         ));
 
         $this->repository->expects($this->once())
             ->method('getPackages')
             ->will($this->returnValue(array()));
 
-        if (!is_dir($this->vendorDir.'/.composer')) {
-            mkdir($this->vendorDir.'/.composer');
+        if (!is_dir($this->vendorDir.'/composer')) {
+            mkdir($this->vendorDir.'/composer');
         }
 
         $this->createClassFile($this->workingDir);
 
-        $this->generator->dump($this->repository, $package, $this->im, $this->vendorDir.'/.composer');
-        $this->assertAutoloadFiles('main', $this->vendorDir.'/.composer');
-        $this->assertAutoloadFiles('classmap', $this->vendorDir.'/.composer', 'classmap');
+        $this->generator->dump($this->repository, $package, $this->im, $this->vendorDir.'/composer');
+        $this->assertAutoloadFiles('main', $this->vendorDir.'/composer');
+        $this->assertAutoloadFiles('classmap', $this->vendorDir.'/composer', 'classmap');
     }
 
     public function testVendorDirSameAsWorkingDir()
@@ -103,22 +103,22 @@ class AutoloadGeneratorTest extends TestCase
         $package = new MemoryPackage('a', '1.0', '1.0');
         $package->setAutoload(array(
             'psr-0' => array('Main' => 'src/', 'Lala' => 'src/'),
-            'classmap' => array('.composersrc/'),
+            'classmap' => array('composersrc/'),
         ));
 
         $this->repository->expects($this->once())
             ->method('getPackages')
             ->will($this->returnValue(array()));
 
-        if (!is_dir($this->vendorDir.'/.composer')) {
-            mkdir($this->vendorDir.'/.composer', 0777, true);
+        if (!is_dir($this->vendorDir.'/composer')) {
+            mkdir($this->vendorDir.'/composer', 0777, true);
         }
 
         $this->createClassFile($this->vendorDir);
 
-        $this->generator->dump($this->repository, $package, $this->im, $this->vendorDir.'/.composer');
-        $this->assertAutoloadFiles('main3', $this->vendorDir.'/.composer');
-        $this->assertAutoloadFiles('classmap3', $this->vendorDir.'/.composer', 'classmap');
+        $this->generator->dump($this->repository, $package, $this->im, $this->vendorDir.'/composer');
+        $this->assertAutoloadFiles('main3', $this->vendorDir.'/composer');
+        $this->assertAutoloadFiles('classmap3', $this->vendorDir.'/composer', 'classmap');
     }
 
     public function testMainPackageAutoloadingAlternativeVendorDir()
@@ -126,7 +126,7 @@ class AutoloadGeneratorTest extends TestCase
         $package = new MemoryPackage('a', '1.0', '1.0');
         $package->setAutoload(array(
             'psr-0' => array('Main' => 'src/', 'Lala' => 'src/'),
-            'classmap' => array('.composersrc/'),
+            'classmap' => array('composersrc/'),
         ));
 
         $this->repository->expects($this->once())
@@ -134,11 +134,11 @@ class AutoloadGeneratorTest extends TestCase
             ->will($this->returnValue(array()));
 
         $this->vendorDir .= '/subdir';
-        mkdir($this->vendorDir.'/.composer', 0777, true);
+        mkdir($this->vendorDir.'/composer', 0777, true);
         $this->createClassFile($this->workingDir);
-        $this->generator->dump($this->repository, $package, $this->im, $this->vendorDir.'/.composer');
-        $this->assertAutoloadFiles('main2', $this->vendorDir.'/.composer');
-        $this->assertAutoloadFiles('classmap2', $this->vendorDir.'/.composer', 'classmap');
+        $this->generator->dump($this->repository, $package, $this->im, $this->vendorDir.'/composer');
+        $this->assertAutoloadFiles('main2', $this->vendorDir.'/composer');
+        $this->assertAutoloadFiles('classmap2', $this->vendorDir.'/composer', 'classmap');
     }
 
     public function testMainPackageAutoloadingWithTargetDir()
@@ -153,8 +153,8 @@ class AutoloadGeneratorTest extends TestCase
             ->method('getPackages')
             ->will($this->returnValue(array()));
 
-        $this->generator->dump($this->repository, $package, $this->im, $this->vendorDir.'/.composer');
-        $this->assertFileEquals(__DIR__.'/Fixtures/autoload_target_dir.php', $this->vendorDir.'/.composer/autoload.php');
+        $this->generator->dump($this->repository, $package, $this->im, $this->vendorDir.'/composer');
+        $this->assertFileEquals(__DIR__.'/Fixtures/autoload_target_dir.php', $this->vendorDir.'/autoload.php');
     }
 
     public function testVendorsAutoloading()
@@ -171,10 +171,10 @@ class AutoloadGeneratorTest extends TestCase
             ->method('getPackages')
             ->will($this->returnValue($packages));
 
-        mkdir($this->vendorDir.'/.composer', 0777, true);
-        $this->generator->dump($this->repository, $package, $this->im, $this->vendorDir.'/.composer');
-        $this->assertAutoloadFiles('vendors', $this->vendorDir.'/.composer');
-        $this->assertTrue(file_exists($this->vendorDir.'/.composer/autoload_classmap.php'), "ClassMap file needs to be generated, even if empty.");
+        mkdir($this->vendorDir.'/composer', 0777, true);
+        $this->generator->dump($this->repository, $package, $this->im, $this->vendorDir.'/composer');
+        $this->assertAutoloadFiles('vendors', $this->vendorDir.'/composer');
+        $this->assertTrue(file_exists($this->vendorDir.'/composer/autoload_classmap.php'), "ClassMap file needs to be generated, even if empty.");
     }
 
     public function testVendorsClassMapAutoloading()
@@ -191,7 +191,7 @@ class AutoloadGeneratorTest extends TestCase
             ->method('getPackages')
             ->will($this->returnValue($packages));
 
-        @mkdir($this->vendorDir.'/.composer', 0777, true);
+        @mkdir($this->vendorDir.'/composer', 0777, true);
         mkdir($this->vendorDir.'/a/a/src', 0777, true);
         mkdir($this->vendorDir.'/b/b/src', 0777, true);
         mkdir($this->vendorDir.'/b/b/lib', 0777, true);
@@ -199,17 +199,17 @@ class AutoloadGeneratorTest extends TestCase
         file_put_contents($this->vendorDir.'/b/b/src/b.php', '<?php class ClassMapBar {}');
         file_put_contents($this->vendorDir.'/b/b/lib/c.php', '<?php class ClassMapBaz {}');
 
-        $this->generator->dump($this->repository, $package, $this->im, $this->vendorDir.'/.composer');
-        $this->assertTrue(file_exists($this->vendorDir.'/.composer/autoload_classmap.php'), "ClassMap file needs to be generated.");
+        $this->generator->dump($this->repository, $package, $this->im, $this->vendorDir.'/composer');
+        $this->assertTrue(file_exists($this->vendorDir.'/composer/autoload_classmap.php'), "ClassMap file needs to be generated.");
         $this->assertEquals(
             array(
                 'ClassMapFoo' => $this->workingDir.'/composer-test-autoload/a/a/src/a.php',
                 'ClassMapBar' => $this->workingDir.'/composer-test-autoload/b/b/src/b.php',
                 'ClassMapBaz' => $this->workingDir.'/composer-test-autoload/b/b/lib/c.php',
             ),
-            include ($this->vendorDir.'/.composer/autoload_classmap.php')
+            include ($this->vendorDir.'/composer/autoload_classmap.php')
         );
-        $this->assertAutoloadFiles('classmap4', $this->vendorDir.'/.composer', 'classmap');
+        $this->assertAutoloadFiles('classmap4', $this->vendorDir.'/composer', 'classmap');
     }
 
     public function testClassMapAutoloadingEmptyDirAndExactFile()
@@ -228,7 +228,7 @@ class AutoloadGeneratorTest extends TestCase
             ->method('getPackages')
             ->will($this->returnValue($packages));
 
-        @mkdir($this->vendorDir.'/.composer', 0777, true);
+        @mkdir($this->vendorDir.'/composer', 0777, true);
         mkdir($this->vendorDir.'/a/a/src', 0777, true);
         mkdir($this->vendorDir.'/b/b', 0777, true);
         mkdir($this->vendorDir.'/c/c/foo', 0777, true);
@@ -236,17 +236,17 @@ class AutoloadGeneratorTest extends TestCase
         file_put_contents($this->vendorDir.'/b/b/test.php', '<?php class ClassMapBar {}');
         file_put_contents($this->vendorDir.'/c/c/foo/test.php', '<?php class ClassMapBaz {}');
 
-        $this->generator->dump($this->repository, $package, $this->im, $this->vendorDir.'/.composer');
-        $this->assertTrue(file_exists($this->vendorDir.'/.composer/autoload_classmap.php'), "ClassMap file needs to be generated.");
+        $this->generator->dump($this->repository, $package, $this->im, $this->vendorDir.'/composer');
+        $this->assertTrue(file_exists($this->vendorDir.'/composer/autoload_classmap.php'), "ClassMap file needs to be generated.");
         $this->assertEquals(
             array(
                 'ClassMapFoo' => $this->workingDir.'/composer-test-autoload/a/a/src/a.php',
                 'ClassMapBar' => $this->workingDir.'/composer-test-autoload/b/b/test.php',
                 'ClassMapBaz' => $this->workingDir.'/composer-test-autoload/c/c/foo/test.php',
             ),
-            include ($this->vendorDir.'/.composer/autoload_classmap.php')
+            include ($this->vendorDir.'/composer/autoload_classmap.php')
         );
-        $this->assertAutoloadFiles('classmap5', $this->vendorDir.'/.composer', 'classmap');
+        $this->assertAutoloadFiles('classmap5', $this->vendorDir.'/composer', 'classmap');
     }
 
     public function testOverrideVendorsAutoloading()
@@ -264,9 +264,9 @@ class AutoloadGeneratorTest extends TestCase
             ->method('getPackages')
             ->will($this->returnValue($packages));
 
-        mkdir($this->vendorDir.'/.composer', 0777, true);
-        $this->generator->dump($this->repository, $package, $this->im, $this->vendorDir.'/.composer');
-        $this->assertAutoloadFiles('override_vendors', $this->vendorDir.'/.composer');
+        mkdir($this->vendorDir.'/composer', 0777, true);
+        $this->generator->dump($this->repository, $package, $this->im, $this->vendorDir.'/composer');
+        $this->assertAutoloadFiles('override_vendors', $this->vendorDir.'/composer');
     }
 
     public function testIncludePathFileGeneration()
@@ -291,18 +291,18 @@ class AutoloadGeneratorTest extends TestCase
             ->method("getPackages")
             ->will($this->returnValue($packages));
 
-        mkdir($this->vendorDir."/.composer", 0777, true);
+        mkdir($this->vendorDir."/composer", 0777, true);
 
-        $this->generator->dump($this->repository, $package, $this->im, $this->vendorDir."/.composer");
+        $this->generator->dump($this->repository, $package, $this->im, $this->vendorDir."/composer");
 
-        $this->assertFileEquals(__DIR__.'/Fixtures/include_paths.php', $this->vendorDir.'/.composer/include_paths.php');
+        $this->assertFileEquals(__DIR__.'/Fixtures/include_paths.php', $this->vendorDir.'/composer/include_paths.php');
         $this->assertEquals(
             array(
                 $this->vendorDir."/a/a/lib",
                 $this->vendorDir."/b/b/library",
                 $this->vendorDir."/c/library",
             ),
-            require($this->vendorDir."/.composer/include_paths.php")
+            require($this->vendorDir."/composer/include_paths.php")
         );
     }
 
@@ -320,13 +320,13 @@ class AutoloadGeneratorTest extends TestCase
             ->method("getPackages")
             ->will($this->returnValue($packages));
 
-        mkdir($this->vendorDir."/.composer", 0777, true);
+        mkdir($this->vendorDir."/composer", 0777, true);
 
-        $this->generator->dump($this->repository, $package, $this->im, $this->vendorDir."/.composer");
+        $this->generator->dump($this->repository, $package, $this->im, $this->vendorDir."/composer");
 
         $oldIncludePath = get_include_path();
 
-        require($this->vendorDir."/.composer/autoload.php");
+        require($this->vendorDir."/autoload.php");
 
         $this->assertEquals(
             $oldIncludePath.PATH_SEPARATOR.$this->vendorDir."/a/a/lib",
@@ -348,20 +348,20 @@ class AutoloadGeneratorTest extends TestCase
             ->method("getPackages")
             ->will($this->returnValue($packages));
 
-        mkdir($this->vendorDir."/.composer", 0777, true);
+        mkdir($this->vendorDir."/composer", 0777, true);
 
-        $this->generator->dump($this->repository, $package, $this->im, $this->vendorDir."/.composer");
+        $this->generator->dump($this->repository, $package, $this->im, $this->vendorDir."/composer");
 
-        $this->assertFalse(file_exists($this->vendorDir."/.composer/include_paths.php"));
+        $this->assertFalse(file_exists($this->vendorDir."/composer/include_paths.php"));
     }
 
     private function createClassFile($basedir)
     {
-        if (!is_dir($basedir.'/.composersrc')) {
-            mkdir($basedir.'/.composersrc', 0777, true);
+        if (!is_dir($basedir.'/composersrc')) {
+            mkdir($basedir.'/composersrc', 0777, true);
         }
 
-        file_put_contents($basedir.'/.composersrc/foo.php', '<?php class ClassMapFoo {}');
+        file_put_contents($basedir.'/composersrc/foo.php', '<?php class ClassMapFoo {}');
     }
 
     private function assertAutoloadFiles($name, $dir, $type = 'namespaces')