+++ CHANGELOG.md
@@ -3,7 +3,7 @@
   * Schema: Added 'require-dev' for development-time requirements (tests, etc), install with --dev
   * Schema: Removed 'recommend'
   * Schema: 'suggest' is now informational and can use any description for a package, not only a constraint
-  * Break: .composer/autoload.php and other files in vendor/.composer have been moved to vendor/
+  * Break: vendor/.composer/autoload.php has been moved to vendor/autoload.php, other files are now in vendor/composer/
   * Added caching of repository metadata (faster startup times & failover if packagist is down)
   * Added include_path support for legacy projects that are full of require_once statements
   * Added installation notifications API to allow better statistics on Composer repositories